---CAMERA (1) Fix Plotting Large Images (2) Get Continuous Mode working in safemode (3) Selecting trigger mode should immediately change the trigger mode on the camera (4) Increment fits increments key file as well. (5) Picture autonormalization option. (6) fix fits files (AGAIN DAMNIT)

+++ Camera/FileSystem.cpp
@@ -108,8 +108,8 @@ int FileSystem::openConfiguration(std::string configurationNameToOpen)
 		int imageBoxWidth = eImageBackgroundAreas[imageLocation].right - eImageBackgroundAreas[imageLocation].left + 1;
 		int imageBoxHeight = eImageBackgroundAreas[imageLocation].bottom - eImageBackgroundAreas[imageLocation].top + 1;
 
-		double boxWidth = imageBoxWidth / eImageWidth;
-		double boxHeight = imageBoxHeight / eImageHeight;
+		double boxWidth = imageBoxWidth / (double)eImageWidth;
+		double boxHeight = imageBoxHeight / (double)eImageHeight;
 		if (boxWidth > boxHeight)
 		{
 			// scale the box width down.
